# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 0.0.1
  title: complex API
host: "localhost:20016"
basePath: /Adam/complex
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
  - text/plain
paths:
...
  /file/upload:
    x-swagger-router-controller: controller
    post:
      description: Uploads the file
      operationId: v1_srvc2006FileUpload
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
      responses:
        "200":
          description: meta data of file
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  "/file/{id}/view":
    x-swagger-router-controller: controller
    get:
      description: View the file
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Id of file
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
      operationId: v1_srvc2006FileView
      responses:
        "200":
          description: file data
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  "/file/download/{id}":
    x-swagger-router-controller: controller
    get:
      description: Download the file
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Id of file
      operationId: v1_srvc2006FileDownload
      responses:
        "200":
          description: file download
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  "/export/download/{id}":
    x-swagger-router-controller: controller
    get:
      description: Download the file
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Id of file
        - name: filename
          in: query
          type: string
          description: filename of file
      operationId: v1_srvc2006ExportedFileDownload
      responses:
        "200":
          description: file download
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  "/fileMapper/{fileId}":
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of 'complex'
      operationId: fileMapperList
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
        - name: fileId
          in: path
          required: true
          type: string
          description: fileId against which we db will be querried
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  "/fileMapper/{fileId}/create":
    x-swagger-router-controller: controller
    post:
      description: Create the data in the file
      operationId: v1_bulkCreate
      parameters:
        - name: data
          in: body
          description: Payload to bulkCreate
          schema:
            $ref: "#/definitions/bulkCreateData"
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - name: fileId
          in: path
          required: true
          type: string
          description: fileId against which we db will be querried
      responses:
        "200":
          description: Stats of bulkCreate
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
...
  /fileMapper/enrich:
    x-swagger-router-controller: controller
    put:
      description: Enrich the sheet data
      operationId: v1_enrichData
      parameters:
        - name: data
          in: body
          description: Payload to validate data
          schema:
            $ref: "#/definitions/enrichData"
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
      responses:
        "200":
          description: Data enriched against mapping
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
...
  "/fileMapper/{fileId}/count":
    x-swagger-router-controller: controller
    get:
      description: returns count of 'complex'
      operationId: fileMapperCount
      parameters:
        - *ref_0
        - *ref_1
        - *ref_2
        - name: fileId
          in: path
          required: true
          type: string
          description: fileId against which we db will be querried
      responses:
        "200":
          description: Count of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
  /export:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of 'complex'
      operationId: v1_srvc2006Export
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
        - *ref_3
        - name: totalRecords
          in: query
          type: integer
          description: total records
      responses:
        "200":
          description: List of the entites
        "400":
          description: Bad parameters
        "500":
          description: Internal server error
...
definitions:
...
  enrichData:
    properties:
      sheetData:
        type:
          - array
        items:
          type: object
      headerMapping:
        type:
          - object
  bulkCreateData:
    properties:
      fileId:
        type: string